change param int to int64

+++ v2/rest/private/orders/cancel-orders.go
@@ -8,8 +8,8 @@ import (
 )
 
 type RequestForCancelOrders struct {
-	Pair     string `json:"pair"`
-	OrderIDs []int  `json:"order_ids"`
+	Pair     string  `json:"pair"`
+	OrderIDs []int64 `json:"order_ids"`
 }
 
 type ResponseForCancelOrders struct {
